Refactor PomodoroCommandManager - rename PomodoroCommandManager to SlackMessagePostedListenerConfiguration - Extract logic to a PomodoCommand factory and each command in its own class
