+++ src/main/java/com/github/jchartier/slackbot/pomodoro/command/PomodoroCommand.java
@@ -1,0 +1,8 @@
+package com.github.jchartier.slackbot.pomodoro.command;
+
+import com.ullink.slack.simpleslackapi.SlackUser;
+
+public interface PomodoroCommand {
+
+    void execute(SlackUser slackUser, String message);
+}
